# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/v1/car-types/": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Create a CarType",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarTypes"
        ],
        "summary": "Create a CarType",
        "parameters": [
          {
            "description": "Create a CarType",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CreateCarTypeRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "CarType response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-types/get-by-filter": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get CarTypes",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarTypes"
        ],
        "summary": "Get CarTypes",
        "parameters": [
          {
            "description": "Request",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CarType response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CarTypeResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-types/{id}": {
      "get": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get a CarType",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarTypes"
        ],
        "summary": "Get a CarType",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "CarType response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Update a CarType",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarTypes"
        ],
        "summary": "Update a CarType",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update a CarType",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCarTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CarType response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Delete a CarType",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarTypes"
        ],
        "summary": "Delete a CarType",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
...
    "/v1/companies/": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Create a Company",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Companies"
        ],
        "summary": "Create a Company",
        "parameters": [
          {
            "description": "Create a Company",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CreateCompanyRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Company response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/companies/get-by-filter": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get Companies",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Companies"
        ],
        "summary": "Get Companies",
        "parameters": [
          {
            "description": "Request",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CompanyResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/companies/{id}": {
      "get": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get a Company",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Companies"
        ],
        "summary": "Get a Company",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Company response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Update a Company",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Companies"
        ],
        "summary": "Update a Company",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update a Company",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCompanyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Delete a Company",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Companies"
        ],
        "summary": "Delete a Company",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CityResponse": {
      "type": "object",
      "properties": {
        "country": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CountryResponse"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse": {
      "type": "object",
      "properties": {
        "country": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CountryResponse"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CountryResponse": {
      "type": "object",
      "properties": {
        "cities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CityResponse"
          }
        },
        "companies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse"
          }
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.CreateCarTypeRequest": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "maxLength": 15,
          "minLength": 3
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.CreateCompanyRequest": {
      "type": "object",
      "required": [
        "countryId",
        "name"
      ],
      "properties": {
        "countryId": {
          "type": "integer"
        },
        "name": {
          "type": "string",
          "maxLength": 20,
          "minLength": 3
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCarTypeRequest": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string",
          "maxLength": 15,
          "minLength": 3
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCompanyRequest": {
      "type": "object",
      "properties": {
        "countryId": {
          "type": "integer"
        },
        "name": {
          "type": "string",
          "maxLength": 20,
          "minLength": 3
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse": {
      "type": "object",
      "properties": {
        "error": {},
        "result": {},
        "resultCode": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.ResultCode"
        },
        "success": {
          "type": "boolean"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_validation.ValidationError"
          }
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_helper.ResultCode": {
      "type": "integer",
      "enum": [
        0,
        40001,
        40101,
        40301,
        40401,
        42901,
        42902,
        50001,
        50002
      ],
      "x-enum-varnames": [
        "Success",
        "ValidationError",
        "AuthError",
        "ForbiddenError",
        "NotFoundError",
        "LimiterError",
        "OtpLimiterError",
        "CustomRecovery",
        "InternalError"
      ]
    },
    "github_com_naeemaei_golang-clean-web-api_api_validation.ValidationError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "property": {
          "type": "string"
        },
        "tag": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_domain_filter.Filter": {
      "type": "object",
      "properties": {
        "filterType": {
          "description": "text number",
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "type": {
          "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CarTypeResponse": {
      "type": "object",
      "properties": {
        "hasNextPage": {
          "type": "boolean"
        },
        "hasPreviousPage": {
          "type": "boolean"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse"
          }
        },
        "pageNumber": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "totalPages": {
          "type": "integer"
        },
        "totalRows": {
          "type": "integer"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CompanyResponse": {
      "type": "object",
      "properties": {
        "hasNextPage": {
          "type": "boolean"
        },
        "hasPreviousPage": {
          "type": "boolean"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse"
          }
        },
        "pageNumber": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "totalPages": {
          "type": "integer"
        },
        "totalRows": {
          "type": "integer"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter": {
      "type": "object",
      "properties": {
        "filter": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.Filter"
          }
        },
        "pageNumber": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "sort": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.Sort"
          }
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_domain_filter.Sort": {
      "type": "object",
      "properties": {
        "colId": {
          "type": "string"
        },
        "sort": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "AuthBearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}